Refactor: useEffect, 버튼핸들러 page단위에 배치
